capture: fix invalid inheritance between Capture classes
Previously, FDCapture inherited from FDCaptureBinary and SysCapture
inherited from SysCaptureBinary. This however is not very proper because
e.g. FDCapture is not a subtype of FDCaptureBinary in the sense that one
returns str and the other bytes. So the subclass is not interchangeable
for its superclass which breaks Liskov substitution. Mypy also complains
about this once type annotations are added.

Instead, have FDCapture *use* a FDCaptureBinary, which doesn't have any
of these issues (same for Sys).