Revert "Move common code between Session and Package to FSCollector"
This reverts commit f10ab021e21a44e2f0fa2be66660c4a6d4b7a61a.

The commit was good in that it removed a non-trivial amount of code
duplication. However it was done in the wrong layer (nodes.py) and split
up a major part of the collection (the filesystem traversal) to a
separate class making it harder to understand.

We should try to reduce the duplication, but in a more appropriate
manner.