mark/structures: better error message for bad `pytestmark`
pytest gets a couple of attributes on objects during collection. For
modules in particular, they are `pytestmark` and `pytest_plugins`. When
a module defines `__getattr__` and doesn't properly take care to raise
`AttributeError` given an unhandled attribute, pytest will raise an
error, but the user will be confused where it came from.

Provide a better error for `pytestmark` when this is detected.
Would have been nice to do this also for `pytest_plugins` but it's not
as easy. Also, `pytest_plugins` allows `None`, which is what you get if
you forget the `AttributeError`, so less likely to be a problem in
practice.

Fix #8265.